+++ src/utils/util.js
@@ -25,3 +25,23 @@ export function isFile(content) {
         content.isNodeJs;
     return (content instanceof File) || isNodeFile;
 }
+
+function computeVersionedHash(commitment, blobCommitmentVersion) {
+    const computedVersionedHash = new Uint8Array(32);
+    computedVersionedHash.set([blobCommitmentVersion], 0);
+    const hash = ethers.getBytes(ethers.sha256(commitment));
+    computedVersionedHash.set(hash.subarray(1), 1);
+    return computedVersionedHash;
+}
+
+export function commitmentsToVersionedHashes(commitment) {
+    return computeVersionedHash(commitment, 0x01);
+}
+
+export function getHash(kzg, blob) {
+    const commit = kzg.blobToKzgCommitment(blob);
+    const localHash = commitmentsToVersionedHashes(commit);
+    const hash = new Uint8Array(32);
+    hash.set(localHash.subarray(0, 32 - 8));
+    return ethers.hexlify(hash);
+}
