+++ src/worker/wasm.js
